[Slider#204] Removed V1, renamed V2
