+++ core/Sources/Components/Slider/UseCase/GetClosestValue/SliderGetClosestValueUseCaseTests.swift
@@ -2,7 +2,7 @@
 //  SliderGetClosestValueUseCaseTests.swift
 //  SparkCoreUnitTests
 //
-//  Created by louis.borlee on 23/11/2023.
+//  Created by louis.borlee on 19/12/2023.
 //  Copyright © 2023 Adevinta. All rights reserved.
 //
 
@@ -11,129 +11,73 @@ import XCTest
 
 final class SliderGetClosestValueUseCaseTests: XCTestCase {
 
-    func test_execute_createValuesFromStep_receivedValues() throws {
+    func test_execute_under_minimum_value() {
         // GIVEN
-        let createValuesFromStepUseCaseMock = SliderCreateValuesFromStepsUseCasableGeneratedMock()
-        createValuesFromStepUseCaseMock._executeWithFromAndToAndSteps = { _, _ , _ throws -> [Float] in
-            return [1, 2, 3]
-        }
-        let sut = SliderGetClosestValueUseCase(createValuesFromStepsUseCase: createValuesFromStepUseCaseMock)
+        let useCase = SliderGetClosestValueUseCase()
+        let values: [CGFloat] = [10, 20, 30]
+        let value: CGFloat = 0
+        let expectedClosestValue: CGFloat = 10
 
         // WHEN
-        _ = sut.execute(from: 2, to: 20, withSteps: 1, fromValue: 15)
+        let closestValue = useCase.execute(value: value, in: values)
 
         // THEN
-        XCTAssertEqual(createValuesFromStepUseCaseMock.executeWithFromAndToAndStepsCallsCount,
-                       1,
-                       "executeWithFromAndToAndStepsCallsCount should be called once")
-        let receivedArguments = try XCTUnwrap(createValuesFromStepUseCaseMock.executeWithFromAndToAndStepsReceivedArguments)
-        XCTAssertEqual(receivedArguments.from, 2, "Wrong receivedArguments.from")
-        XCTAssertEqual(receivedArguments.to, 20, "Wrong receivedArguments.to")
-        XCTAssertEqual(receivedArguments.steps, 1, "Wrong receivedArguments.steps")
+        XCTAssertEqual(closestValue, expectedClosestValue)
     }
 
-    func test_execute_createValuesFromStep_throws_invalidRange() {
+    func test_execute_over_maximum_value() {
         // GIVEN
-        let createValuesFromStepUseCaseMock = SliderCreateValuesFromStepsUseCasableGeneratedMock()
-        createValuesFromStepUseCaseMock._executeWithFromAndToAndSteps = { _, _ , _ throws -> [Float] in
-            throw SliderCreateValuesFromStepsUseCasableError.invalidRange
-        }
-        let sut = SliderGetClosestValueUseCase(createValuesFromStepsUseCase: createValuesFromStepUseCaseMock)
+        let useCase = SliderGetClosestValueUseCase()
+        let values: [Double] = [0.1, 0.2, 0.3]
+        let value: Double = 0.4
+        let expectedClosestValue: Double = 0.3
 
         // WHEN
-        let value = sut.execute(from: 0, to: 100, withSteps: 10, fromValue: 40)
+        let closestValue = useCase.execute(value: value, in: values)
 
         // THEN
-        XCTAssertEqual(value, 40)
+        XCTAssertEqual(closestValue, expectedClosestValue)
     }
 
-    func test_execute_createValuesFromStep_throws_invalid() {
+    func test_execute_lower_rounding() {
         // GIVEN
-        let createValuesFromStepUseCaseMock = SliderCreateValuesFromStepsUseCasableGeneratedMock()
-        createValuesFromStepUseCaseMock._executeWithFromAndToAndSteps = { _, _ , _ throws -> [Float] in
-            throw SliderCreateValuesFromStepsUseCasableError.invalidStep
-        }
-        let sut = SliderGetClosestValueUseCase(createValuesFromStepsUseCase: createValuesFromStepUseCaseMock)
+        let useCase = SliderGetClosestValueUseCase()
+        let values: [Float] = [0.0, 0.50, 1.0, 1.50]
+        let value: Float = 0.74
+        let expectedClosestValue: Float = 0.50
 
         // WHEN
-        let value = sut.execute(from: 0, to: 100, withSteps: 10, fromValue: 30)
+        let closestValue = useCase.execute(value: value, in: values)
 
         // THEN
-        XCTAssertEqual(value, 30)
+        XCTAssertEqual(closestValue, expectedClosestValue)
     }
 
-    func test_execute_createValuesFromStep_empty_values() {
+    func test_execute_upper_rounding() {
         // GIVEN
-        let createValuesFromStepUseCaseMock = SliderCreateValuesFromStepsUseCasableGeneratedMock()
-        createValuesFromStepUseCaseMock._executeWithFromAndToAndSteps = { _, _ , _ throws -> [Float] in
-            return []
-        }
-        let sut = SliderGetClosestValueUseCase(createValuesFromStepsUseCase: createValuesFromStepUseCaseMock)
+        let useCase = SliderGetClosestValueUseCase()
+        let values: [CGFloat] = [0.0, 0.50, 1.0, 1.50]
+        let value: CGFloat = 0.76
+        let expectedClosestValue: CGFloat = 1.0
 
         // WHEN
-        let value = sut.execute(from: 0, to: 100, withSteps: 10, fromValue: 60)
+        let closestValue = useCase.execute(value: value, in: values)
 
         // THEN
-        XCTAssertEqual(value, 60)
+        XCTAssertEqual(closestValue, expectedClosestValue)
     }
 
     func test_execute_inbetween() {
         // GIVEN
-        let createValuesFromStepUseCaseMock = SliderCreateValuesFromStepsUseCasableGeneratedMock()
-        createValuesFromStepUseCaseMock._executeWithFromAndToAndSteps = { _, _ , _ throws -> [Float] in
-            return [
-                0,
-                10,
-                20,
-                30,
-                40,
-                50
-            ]
-        }
-        let sut = SliderGetClosestValueUseCase(createValuesFromStepsUseCase: createValuesFromStepUseCaseMock)
+        let useCase = SliderGetClosestValueUseCase()
+        let values: [CGFloat] = [0.0, 0.50, 1.0, 1.50]
+        let value: CGFloat = 0.75
+        let expectedClosestValue: CGFloat = 1.0
 
         // WHEN
-        let value = sut.execute(from: 0, to: 50, withSteps: 0.1, fromValue: 25)
+        let closestValue = useCase.execute(value: value, in: values)
 
         // THEN
-        XCTAssertEqual(value, 30)
-    }
-
-    func test_execute_lower() {
-        // GIVEN
-        let createValuesFromStepUseCaseMock = SliderCreateValuesFromStepsUseCasableGeneratedMock()
-        createValuesFromStepUseCaseMock._executeWithFromAndToAndSteps = { _, _ , _ throws -> [Float] in
-            return [
-                0.0,
-                0.5
-            ]
-        }
-        let sut = SliderGetClosestValueUseCase(createValuesFromStepsUseCase: createValuesFromStepUseCaseMock)
-
-        // WHEN
-        let value = sut.execute(from: 0, to: 0.5, withSteps: 0.1, fromValue: 0.249)
-
-        // THEN
-        XCTAssertEqual(value, 0.0)
-    }
-
-    func test_execute_upper() {
-        // GIVEN
-        let createValuesFromStepUseCaseMock = SliderCreateValuesFromStepsUseCasableGeneratedMock()
-        createValuesFromStepUseCaseMock._executeWithFromAndToAndSteps = { _, _ , _ throws -> [Float] in
-            return [
-                0.0,
-                0.1,
-                0.4,
-                0.5
-            ]
-        }
-        let sut = SliderGetClosestValueUseCase(createValuesFromStepsUseCase: createValuesFromStepUseCaseMock)
-
-        // WHEN
-        let value = sut.execute(from: 0, to: 0.5, withSteps: 0.1, fromValue: 0.251)
-
-        // THEN
-        XCTAssertEqual(value, 0.4)
+        XCTAssertEqual(closestValue, expectedClosestValue)
     }
 }
