+++ core/Sources/Components/Slider/ViewModel/Base/SliderViewModelTests.swift
@@ -2,56 +2,28 @@
 //  SliderViewModelTests.swift
 //  SparkCoreUnitTests
 //
-//  Created by louis.borlee on 06/12/2023.
-//  Copyright © 2023 Adevinta. All rights reserved.
+//  Created by louis.borlee on 02/01/2024.
+//  Copyright © 2024 Adevinta. All rights reserved.
 //
 
 import XCTest
 import Combine
 @testable import SparkCore
 
-class SliderViewModelTests<T: SliderViewModel>: XCTestCase {
-    private let intent = SliderIntent.info
-    private let shape = SliderShape.rounded
-    private let expectedRadii = SliderRadii.mocked()
-
-    private var theme: ThemeGeneratedMock!
-    var viewModel: T!
-    private var expectedColors: SliderColors!
-    var getColorsUseCase: SliderGetColorsUseCasableGeneratedMock!
-    var getCornerRadiiUseCase: SliderGetCornerRadiiUseCasableGeneratedMock!
-    var getClosestValueUseCase: SliderGetClosestValueUseCasableGeneratedMock!
-    var publishers: SliderPublishers!
+final class SliderViewModelTests: SliderViewModelWithMocksTests {
 
     override func setUp() {
         super.setUp()
-        self.theme = ThemeGeneratedMock.mocked()
-        self.expectedColors = SliderColors.mocked(colors: self.theme.colors)
-        self.getColorsUseCase = SliderGetColorsUseCasableGeneratedMock.mocked(returnedColors: self.expectedColors)
-        self.getCornerRadiiUseCase = SliderGetCornerRadiiUseCasableGeneratedMock.mocked(expectedRadii: self.expectedRadii)
-        self.getClosestValueUseCase = SliderGetClosestValueUseCasableGeneratedMock()
-        self.viewModel = T(
+        self.viewModel = SliderViewModel(
             theme: self.theme,
             shape: self.shape,
             intent: self.intent,
             getColorsUseCase: self.getColorsUseCase,
             getCornerRadiiUseCase: self.getCornerRadiiUseCase,
+            getStepValuesInBoundsUseCase: self.getStepValuesInBoundsUseCase,
             getClosestValueUseCase: self.getClosestValueUseCase
         )
         self.setupPublishers()
-    }
-
-    func setupPublishers() {
-        self.publishers = SliderPublishers(
-            dim: PublisherMock(publisher: self.viewModel.$dim),
-            trackColor: PublisherMock(publisher: self.viewModel.$trackColor),
-            handleColor: PublisherMock(publisher: self.viewModel.$handleColor),
-            indicatorColor: PublisherMock(publisher: self.viewModel.$indicatorColor),
-            handleActiveIndicatorColor: PublisherMock(publisher: self.viewModel.$handleActiveIndicatorColor),
-            trackRadius: PublisherMock(publisher: self.viewModel.$trackRadius),
-            indicatorRadius: PublisherMock(publisher: self.viewModel.$indicatorRadius)
-        )
-        self.publishers.load()
     }
 
     // MARK: - init
@@ -62,6 +34,8 @@ class SliderViewModelTests<T: SliderViewModel>: XCTestCase {
         XCTAssertEqual(self.viewModel.intent, self.intent, "Wrong theme")
         XCTAssertEqual(self.viewModel.shape, self.shape, "Wrong shape")
         XCTAssertEqual(self.viewModel.dim, self.theme.dims.none, "Wrong dim")
+        XCTAssertNil(self.viewModel.step, "step should be nil")
+        XCTAssertNil(self.viewModel.stepValues, "stepValues should be nil")
 
         // THEN - Corner Radii
         XCTAssertEqual(self.getCornerRadiiUseCase.executeWithThemeAndShapeCallsCount, 1, "getCornerRadiiUseCase.executeWithThemeAndShape should be called once")
@@ -81,8 +55,11 @@ class SliderViewModelTests<T: SliderViewModel>: XCTestCase {
         XCTAssertEqual(self.viewModel.handleColor.uiColor, self.expectedColors.handle.uiColor, "Wrong handleColor")
         XCTAssertEqual(self.viewModel.handleActiveIndicatorColor.uiColor, self.expectedColors.handleActiveIndicator.uiColor, "Wrong handleActiveIndicatorColor")
 
+        // THEN - StepValuesInBounds
+        XCTAssertFalse(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepCalled, "getStepValuesInBoundsUseCase.executeWithBoundsAndStep shouldn't have been called")
+
         // THEN - ClosestValues
-        XCTAssertFalse(self.getClosestValueUseCase.executeWithFromAndToAndStepsAndValueCalled, "getClosestValueUseCase.executeWithFromAndToAndStepsAndValue shouldn't have been called")
+        XCTAssertFalse(self.getClosestValueUseCase.executeWithValueAndValuesCalled, "getClosestValueUseCase.executeWithValueAndValues shouldn't have been called")
 
         // THEN - Publishers
         XCTAssertEqual(self.publishers.dim.sinkCount, 1, "$dim should have been called once")
@@ -119,8 +96,11 @@ class SliderViewModelTests<T: SliderViewModel>: XCTestCase {
         let colorsReceivedArguments = try XCTUnwrap(self.getColorsUseCase.executeWithThemeAndIntentReceivedArguments, "Couldn't unwrap colorsReceivedArguments")
         XCTAssertIdentical(colorsReceivedArguments.theme as? ThemeGeneratedMock, newTheme, "Wrong colorsReceivedArguments.theme")
 
+        // THEN - StepValuesInBounds
+        XCTAssertFalse(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepCalled, "getStepValuesInBoundsUseCase.executeWithBoundsAndStep shouldn't have been called")
+
         // THEN - ClosestValues
-        XCTAssertFalse(self.getClosestValueUseCase.executeWithFromAndToAndStepsAndValueCalled, "getClosestValueUseCase.executeWithFromAndToAndStepsAndValue shouldn't have been called")
+        XCTAssertFalse(self.getClosestValueUseCase.executeWithValueAndValuesCalled, "getClosestValueUseCase.executeWithValueAndValues shouldn't have been called")
 
         XCTAssertEqual(self.publishers.dim.sinkCount, 1, "$dim should have been called once")
         XCTAssertEqual(self.publishers.handleColor.sinkCount, 1, "$handleColor should have been called once")
@@ -143,16 +123,17 @@ class SliderViewModelTests<T: SliderViewModel>: XCTestCase {
         // THEN - UseCases
         XCTAssertFalse(self.getColorsUseCase.executeWithThemeAndIntentCalled, "getColorsUseCase.executeWithThemeAndIntent shouldn't have been called")
         XCTAssertFalse(self.getCornerRadiiUseCase.executeWithThemeAndShapeCalled, "getCornerRadiiUseCase.executeWithThemeAndShape shouldn't have been called")
-        XCTAssertFalse(self.getClosestValueUseCase.executeWithFromAndToAndStepsAndValueCalled, "getClosestValueUseCase.executeWithFromAndToAndStepsAndValue shouldn't have been called")
+        XCTAssertFalse(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepCalled, "getStepValuesInBoundsUseCase.executeWithBoundsAndStep shouldn't have been called")
+        XCTAssertFalse(self.getClosestValueUseCase.executeWithValueAndValuesCalled, "getClosestValueUseCase.executeWithValueAndValues shouldn't have been called")
 
         // THEN - Publishers
         XCTAssertEqual(self.publishers.dim.sinkCount, 1, "$dim should be called once")
-        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor should not have been called")
-        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor should not have been called")
-        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor should not have been called")
-        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor should not have been called")
-        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius should not have been called")
-        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius should not have been called")
+        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius shouldn't have been called")
     }
 
     func test_isEnabled_didSet_equal() {
@@ -166,16 +147,17 @@ class SliderViewModelTests<T: SliderViewModel>: XCTestCase {
         // THEN - UseCases
         XCTAssertFalse(self.getColorsUseCase.executeWithThemeAndIntentCalled, "getColorsUseCase.executeWithThemeAndIntent shouldn't have been called")
         XCTAssertFalse(self.getCornerRadiiUseCase.executeWithThemeAndShapeCalled, "getCornerRadiiUseCase.executeWithThemeAndShape shouldn't have been called")
-        XCTAssertFalse(self.getClosestValueUseCase.executeWithFromAndToAndStepsAndValueCalled, "getClosestValueUseCase.executeWithFromAndToAndStepsAndValue shouldn't have been called")
+        XCTAssertFalse(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepCalled, "getStepValuesInBoundsUseCase.executeWithBoundsAndStep shouldn't have been called")
+        XCTAssertFalse(self.getClosestValueUseCase.executeWithValueAndValuesCalled, "getClosestValueUseCase.executeWithValueAndValues shouldn't have been called")
 
         // THEN - Publishers
-        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim should not have been called")
-        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor should not have been called")
-        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor should not have been called")
-        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor should not have been called")
-        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor should not have been called")
-        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius should not have been called")
-        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius should not have been called")
+        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius shouldn't have been called")
     }
 
     // MARK: - Intent
@@ -190,16 +172,17 @@ class SliderViewModelTests<T: SliderViewModel>: XCTestCase {
         // THEN - UseCases
         XCTAssertEqual(self.getColorsUseCase.executeWithThemeAndIntentCallsCount, 1, "getColorsUseCase.executeWithThemeAndIntent should have been called once")
         XCTAssertFalse(self.getCornerRadiiUseCase.executeWithThemeAndShapeCalled, "getCornerRadiiUseCase.executeWithThemeAndShape shouldn't have been called")
-        XCTAssertFalse(self.getClosestValueUseCase.executeWithFromAndToAndStepsAndValueCalled, "getClosestValueUseCase.executeWithFromAndToAndStepsAndValue shouldn't have been called")
+        XCTAssertFalse(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepCalled, "getStepValuesInBoundsUseCase.executeWithBoundsAndStep shouldn't have been called")
+        XCTAssertFalse(self.getClosestValueUseCase.executeWithValueAndValuesCalled, "getClosestValueUseCase.executeWithValueAndValues shouldn't have been called")
 
         // THEN - Publishers
-        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim should not have been called")
+        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim shouldn't have been called")
         XCTAssertEqual(self.publishers.handleColor.sinkCount, 1, "$handleColor should have been called once")
         XCTAssertEqual(self.publishers.handleActiveIndicatorColor.sinkCount, 1, "$handleActiveIndicatorColor should have been called once")
         XCTAssertEqual(self.publishers.trackColor.sinkCount, 1, "$trackColor should have been called once")
         XCTAssertEqual(self.publishers.indicatorColor.sinkCount, 1, "$indicatorColor should have been called once")
-        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius should not have been called")
-        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius should not have been called")
+        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius shouldn't have been called")
     }
 
     func test_intent_didSet_equal() {
@@ -213,16 +196,17 @@ class SliderViewModelTests<T: SliderViewModel>: XCTestCase {
         // THEN - UseCases
         XCTAssertFalse(self.getColorsUseCase.executeWithThemeAndIntentCalled, "getColorsUseCase.executeWithThemeAndIntent shouldn't have been called")
         XCTAssertFalse(self.getCornerRadiiUseCase.executeWithThemeAndShapeCalled, "getCornerRadiiUseCase.executeWithThemeAndShape shouldn't have been called")
-        XCTAssertFalse(self.getClosestValueUseCase.executeWithFromAndToAndStepsAndValueCalled, "getClosestValueUseCase.executeWithFromAndToAndStepsAndValue shouldn't have been called")
+        XCTAssertFalse(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepCalled, "getStepValuesInBoundsUseCase.executeWithBoundsAndStep shouldn't have been called")
+        XCTAssertFalse(self.getClosestValueUseCase.executeWithValueAndValuesCalled, "getClosestValueUseCase.executeWithValueAndValues shouldn't have been called")
 
         // THEN - Publishers
-        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim should not have been called")
-        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor should not have been called")
-        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor should not have been called")
-        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor should not have been called")
-        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor should not have been called")
-        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius should not have been called")
-        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius should not have been called")
+        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius shouldn't have been called")
     }
 
     // MARK: - Shape
@@ -237,14 +221,15 @@ class SliderViewModelTests<T: SliderViewModel>: XCTestCase {
         // THEN - UseCases
         XCTAssertFalse(self.getColorsUseCase.executeWithThemeAndIntentCalled, "getColorsUseCase.executeWithThemeAndIntent shouldn't have been called")
         XCTAssertEqual(self.getCornerRadiiUseCase.executeWithThemeAndShapeCallsCount, 1, "getCornerRadiiUseCase.executeWithThemeAndShape should have been called once")
-        XCTAssertFalse(self.getClosestValueUseCase.executeWithFromAndToAndStepsAndValueCalled, "getClosestValueUseCase.executeWithFromAndToAndStepsAndValue shouldn't have been called")
+        XCTAssertFalse(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepCalled, "getStepValuesInBoundsUseCase.executeWithBoundsAndStep shouldn't have been called")
+        XCTAssertFalse(self.getClosestValueUseCase.executeWithValueAndValuesCalled, "getClosestValueUseCase.executeWithValueAndValues shouldn't have been called")
 
         // THEN - Publishers
-        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim should not have been called")
-        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor should not have been called")
-        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor should not have been called")
-        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor should not have been called")
-        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor should not have been called")
+        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor shouldn't have been called")
         XCTAssertEqual(self.publishers.trackRadius.sinkCount, 1, "$trackRadius should have been called once")
         XCTAssertEqual(self.publishers.indicatorRadius.sinkCount, 1, "$indicatorRadius should have been called once")
     }
@@ -260,16 +245,168 @@ class SliderViewModelTests<T: SliderViewModel>: XCTestCase {
         // THEN - UseCases
         XCTAssertFalse(self.getColorsUseCase.executeWithThemeAndIntentCalled, "getColorsUseCase.executeWithThemeAndIntent shouldn't have been called")
         XCTAssertFalse(self.getCornerRadiiUseCase.executeWithThemeAndShapeCalled, "getCornerRadiiUseCase.executeWithThemeAndShape shouldn't have been called")
-        XCTAssertFalse(self.getClosestValueUseCase.executeWithFromAndToAndStepsAndValueCalled, "getClosestValueUseCase.executeWithFromAndToAndStepsAndValue shouldn't have been called")
+        XCTAssertFalse(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepCalled, "getStepValuesInBoundsUseCase.executeWithBoundsAndStep shouldn't have been called")
+        XCTAssertFalse(self.getClosestValueUseCase.executeWithValueAndValuesCalled, "getClosestValueUseCase.executeWithValueAndValues shouldn't have been called")
 
         // THEN - Publishers
-        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim should not have been called")
-        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor should not have been called")
-        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor should not have been called")
-        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor should not have been called")
-        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor should not have been called")
-        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius should not have been called")
-        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius should not have been called")
+        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius shouldn't have been called")
+    }
+
+    // MARK: - Step
+    func test_step_same_value() {
+        // GIVEN - Inits from setUp()
+        self.resetUseCases() // Removes execute from init
+        self.publishers.reset() // Removes publishes from init
+
+        // WHEN
+        self.viewModel.step = nil
+
+        // THEN
+        XCTAssertNil(self.viewModel.stepValues, "stepValues should be nil")
+
+        // THEN - UseCases
+        XCTAssertFalse(self.getColorsUseCase.executeWithThemeAndIntentCalled, "getColorsUseCase.executeWithThemeAndIntent shouldn't have been called")
+        XCTAssertFalse(self.getCornerRadiiUseCase.executeWithThemeAndShapeCalled, "getCornerRadiiUseCase.executeWithThemeAndShape shouldn't have been called")
+        XCTAssertFalse(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepCalled, "getStepValuesInBoundsUseCase.executeWithBoundsAndStep shouldn't have been called")
+        XCTAssertFalse(self.getClosestValueUseCase.executeWithValueAndValuesCalled, "getClosestValueUseCase.executeWithValueAndValues shouldn't have been called")
+
+        // THEN - Publishers
+        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius shouldn't have been called")
+    }
+
+    func test_step_new_value() throws {
+        // GIVEN - Inits from setUp()
+        let expectedStepValues: [Float] = [0, 1, 2]
+        self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepReturnValue = expectedStepValues
+        self.resetUseCases() // Removes execute from init
+        self.publishers.reset() // Removes publishes from init
+
+        // WHEN
+        self.viewModel.step = 0.3
+
+        // THEN
+        XCTAssertEqual(self.viewModel.stepValues, expectedStepValues, "Wrong stepValues")
+
+        // THEN - UseCases
+        XCTAssertFalse(self.getColorsUseCase.executeWithThemeAndIntentCalled, "getColorsUseCase.executeWithThemeAndIntent shouldn't have been called")
+        XCTAssertFalse(self.getCornerRadiiUseCase.executeWithThemeAndShapeCalled, "getCornerRadiiUseCase.executeWithThemeAndShape shouldn't have been called")
+        XCTAssertFalse(self.getClosestValueUseCase.executeWithValueAndValuesCalled, "getClosestValueUseCase.executeWithValueAndValues shouldn't have been called")
+        XCTAssertEqual(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepCallsCount, 1, "getStepValuesInBoundsUseCase.executeWithBoundsAndStep shouldn have been called")
+        let receivedArguments = try XCTUnwrap(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepReceivedArguments, "Couldn't unwrap receivedArguments")
+        XCTAssertEqual(receivedArguments.step, 0.3, "Wrong received step")
+        XCTAssertEqual(receivedArguments.bounds, 0...1, "Wrong received bounds")
+
+        // THEN - Publishers
+        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius shouldn't have been called")
+    }
+
+    func test_step_nil_value() throws {
+        // GIVEN - Inits from setUp()
+        let expectedStepValues: [Float] = [0, 1, 2]
+        self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepReturnValue = expectedStepValues
+        self.viewModel.step = 0.3
+        self.resetUseCases() // Removes execute from init
+        self.publishers.reset() // Removes publishes from init
+
+        // WHEN
+        self.viewModel.step = nil
+
+        // THEN
+        XCTAssertNil(self.viewModel.stepValues, "stepValues should be nil")
+
+        // THEN - UseCases
+        XCTAssertFalse(self.getColorsUseCase.executeWithThemeAndIntentCalled, "getColorsUseCase.executeWithThemeAndIntent shouldn't have been called")
+        XCTAssertFalse(self.getCornerRadiiUseCase.executeWithThemeAndShapeCalled, "getCornerRadiiUseCase.executeWithThemeAndShape shouldn't have been called")
+        XCTAssertFalse(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepCalled, "getStepValuesInBoundsUseCase.executeWithBoundsAndStep shouldn't have been called")
+        XCTAssertFalse(self.getClosestValueUseCase.executeWithValueAndValuesCalled, "getClosestValueUseCase.executeWithValueAndValues shouldn't have been called")
+
+        // THEN - Publishers
+        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius shouldn't have been called")
+    }
+
+    // MARK: - Bounds
+    func test_bounds_same_value() {
+        // GIVEN - Inits from setUp()
+        self.resetUseCases() // Removes execute from init
+        self.publishers.reset() // Removes publishes from init
+
+        // WHEN
+        self.viewModel.bounds = 0...1
+
+        // THEN
+        XCTAssertNil(self.viewModel.stepValues, "stepValues should be nil")
+
+        // THEN - UseCases
+        XCTAssertFalse(self.getColorsUseCase.executeWithThemeAndIntentCalled, "getColorsUseCase.executeWithThemeAndIntent shouldn't have been called")
+        XCTAssertFalse(self.getCornerRadiiUseCase.executeWithThemeAndShapeCalled, "getCornerRadiiUseCase.executeWithThemeAndShape shouldn't have been called")
+        XCTAssertFalse(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepCalled, "getStepValuesInBoundsUseCase.executeWithBoundsAndStep shouldn't have been called")
+        XCTAssertFalse(self.getClosestValueUseCase.executeWithValueAndValuesCalled, "getClosestValueUseCase.executeWithValueAndValues shouldn't have been called")
+
+        // THEN - Publishers
+        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius shouldn't have been called")
+    }
+
+    func test_bounds_new_value() throws {
+        // GIVEN - Inits from setUp()
+        let expectedStepValues: [Float] = [0, 1, 2]
+        self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepReturnValue = expectedStepValues
+        self.viewModel.step = 0.4 // Changing bounds will not change stepValues unless step isn't nil
+        self.resetUseCases() // Removes previous executes
+        self.publishers.reset() // Removes previous publishes
+
+        // WHEN
+        self.viewModel.bounds = 0...4
+
+        // THEN
+        XCTAssertEqual(self.viewModel.stepValues, expectedStepValues, "Wrong stepValues")
+
+        // THEN - UseCases
+        XCTAssertFalse(self.getColorsUseCase.executeWithThemeAndIntentCalled, "getColorsUseCase.executeWithThemeAndIntent shouldn't have been called")
+        XCTAssertFalse(self.getCornerRadiiUseCase.executeWithThemeAndShapeCalled, "getCornerRadiiUseCase.executeWithThemeAndShape shouldn't have been called")
+        XCTAssertFalse(self.getClosestValueUseCase.executeWithValueAndValuesCalled, "getClosestValueUseCase.executeWithValueAndValues shouldn't have been called")
+        XCTAssertEqual(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepCallsCount, 1, "getStepValuesInBoundsUseCase.executeWithBoundsAndStep shouldn have been called")
+        let receivedArguments = try XCTUnwrap(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepReceivedArguments, "Couldn't unwrap receivedArguments")
+        XCTAssertEqual(receivedArguments.step, 0.4, "Wrong received step")
+        XCTAssertEqual(receivedArguments.bounds, 0...4, "Wrong received bounds")
+
+        // THEN - Publishers
+        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius shouldn't have been called")
     }
 
     // MARK: - Get Closest Value
@@ -277,73 +414,182 @@ class SliderViewModelTests<T: SliderViewModel>: XCTestCase {
         // GIVEN - Inits from setUp()
         self.resetUseCases() // Removes execute from init
         self.publishers.reset() // Removes publishes from init
-        self.getClosestValueUseCase.executeWithFromAndToAndStepsAndValueReturnValue = 0.31
 
         // WHEN
         let returnedValue = self.viewModel.getClosestValue(fromValue: 0.4)
 
         // THEN
-        XCTAssertEqual(returnedValue, 0.31, "Wrong returnedValue")
+        XCTAssertEqual(returnedValue, 0.4, "Wrong returnedValue")
 
         // THEN - UseCases
         XCTAssertFalse(self.getColorsUseCase.executeWithThemeAndIntentCalled, "getColorsUseCase.executeWithThemeAndIntent shouldn't have been called")
-        XCTAssertFalse(self.getCornerRadiiUseCase.executeWithThemeAndShapeCalled, "getCornerRadiiUseCase.executeWithThemeAndShape should not have been called")
-        XCTAssertEqual(self.getClosestValueUseCase.executeWithFromAndToAndStepsAndValueCallsCount, 1, "getClosestValueUseCase.executeWithFromAndToAndStepsAndValue should be called once")
-        let receivedArguments = try XCTUnwrap(self.getClosestValueUseCase.executeWithFromAndToAndStepsAndValueReceivedArguments, "Couldn't unwrap receivedArguments")
-        XCTAssertEqual(receivedArguments.from, 0, "Wrong received from")
-        XCTAssertEqual(receivedArguments.to, 1, "Wrong received to")
-        XCTAssertEqual(receivedArguments.steps, 0.0, "Wrong received steps")
-        XCTAssertEqual(receivedArguments.value, 0.4, "Wrong received value")
+        XCTAssertFalse(self.getCornerRadiiUseCase.executeWithThemeAndShapeCalled, "getCornerRadiiUseCase.executeWithThemeAndShape shouldn't have been called")
+        XCTAssertFalse(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepCalled, "getStepValuesInBoundsUseCase.executeWithBoundsAndStep shouldn't have been called")
+        XCTAssertFalse(self.getClosestValueUseCase.executeWithValueAndValuesCalled, "getClosestValueUseCase.executeWithValueAndValues shouldn't have been called")
 
         // THEN - Publishers
-        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim should not have been called")
-        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor should not have been called")
-        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor should not have been called")
-        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor should not have been called")
-        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor should not have been called")
-        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius should not have been called")
-        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius should not have been called")
+        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius shouldn't have been called")
     }
 
-    func test_getClosestValue_updated_variables() throws {
+    func test_getClosestValue_custom_variables() throws {
+        // GIVEN - Inits from setUp()
+        let stepValues: [Float] = [0, 3, 6, 9]
+        self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepReturnValue = stepValues
+        self.getClosestValueUseCase.executeWithValueAndValuesReturnValue = 6
+        self.viewModel.step = 3
+        self.viewModel.bounds = 0...10
+        self.resetUseCases() // Removes previous executes
+        self.publishers.reset() // Removes previous publishes
+
+        // WHEN
+        let returnedValue = self.viewModel.getClosestValue(fromValue: 5)
+
+        // THEN
+        XCTAssertEqual(returnedValue, 6.0, "Wrong returnedValue")
+
+        // THEN - UseCases
+        XCTAssertFalse(self.getColorsUseCase.executeWithThemeAndIntentCalled, "getColorsUseCase.executeWithThemeAndIntent shouldn't have been called")
+        XCTAssertFalse(self.getCornerRadiiUseCase.executeWithThemeAndShapeCalled, "getCornerRadiiUseCase.executeWithThemeAndShape shouldn't have been called")
+        XCTAssertFalse(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepCalled, "getStepValuesInBoundsUseCase.executeWithBoundsAndStep shouldn't have been called")
+        XCTAssertEqual(self.getClosestValueUseCase.executeWithValueAndValuesCallsCount, 1, "getClosestValueUseCase.executeWithValueAndValues should have been called once")
+        let receivedArguments = try XCTUnwrap(self.getClosestValueUseCase.executeWithValueAndValuesReceivedArguments, "Couldn't unwrap receivedArguments")
+        XCTAssertEqual(receivedArguments.value, 5, "Wrong received value")
+        XCTAssertEqual(receivedArguments.values, stepValues, "Wrong received value")
+
+        // THEN - Publishers
+        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius shouldn't have been called")
+    }
+
+    func test_getClosestValue_value_over_bounds() throws {
         // GIVEN - Inits from setUp()
         self.resetUseCases() // Removes execute from init
         self.publishers.reset() // Removes publishes from init
-        self.getClosestValueUseCase.executeWithFromAndToAndStepsAndValueReturnValue = 7.6
-        self.viewModel.steps = 0.5
-        self.viewModel.minimumValue = 1
-        self.viewModel.maximumValue = 10
 
         // WHEN
-        let returnedValue = self.viewModel.getClosestValue(fromValue: 4.23)
+        let returnedValue = self.viewModel.getClosestValue(fromValue: 2)
 
         // THEN
-        XCTAssertEqual(returnedValue, 7.6, "Wrong returnedValue")
+        XCTAssertEqual(returnedValue, 1, "Wrong returnedValue")
 
         // THEN - UseCases
         XCTAssertFalse(self.getColorsUseCase.executeWithThemeAndIntentCalled, "getColorsUseCase.executeWithThemeAndIntent shouldn't have been called")
-        XCTAssertFalse(self.getCornerRadiiUseCase.executeWithThemeAndShapeCalled, "getCornerRadiiUseCase.executeWithThemeAndShape should not have been called")
-        XCTAssertEqual(self.getClosestValueUseCase.executeWithFromAndToAndStepsAndValueCallsCount, 1, "getClosestValueUseCase.executeWithFromAndToAndStepsAndValue should be called once")
-        let receivedArguments = try XCTUnwrap(self.getClosestValueUseCase.executeWithFromAndToAndStepsAndValueReceivedArguments, "Couldn't unwrap receivedArguments")
-        XCTAssertEqual(receivedArguments.from, 1, "Wrong received from")
-        XCTAssertEqual(receivedArguments.to, 10, "Wrong received to")
-        XCTAssertEqual(receivedArguments.steps, 0.5, "Wrong received steps")
-        XCTAssertEqual(receivedArguments.value, 4.23, "Wrong received value")
+        XCTAssertFalse(self.getCornerRadiiUseCase.executeWithThemeAndShapeCalled, "getCornerRadiiUseCase.executeWithThemeAndShape shouldn't have been called")
+        XCTAssertFalse(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepCalled, "getStepValuesInBoundsUseCase.executeWithBoundsAndStep shouldn't have been called")
+        XCTAssertFalse(self.getClosestValueUseCase.executeWithValueAndValuesCalled, "getClosestValueUseCase.executeWithValueAndValues shouldn't have been called")
 
         // THEN - Publishers
-        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim should not have been called")
-        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor should not have been called")
-        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor should not have been called")
-        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor should not have been called")
-        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor should not have been called")
-        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius should not have been called")
-        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius should not have been called")
+        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius shouldn't have been called")
     }
 
-    func resetUseCases() {
-        self.getColorsUseCase.reset()
-        self.getCornerRadiiUseCase.reset()
-        self.getClosestValueUseCase.reset()
+    func test_getClosestValue_value_under_bounds() throws {
+        // GIVEN - Inits from setUp()
+        self.resetUseCases() // Removes execute from init
+        self.publishers.reset() // Removes publishes from init
+
+        // WHEN
+        let returnedValue = self.viewModel.getClosestValue(fromValue: -1)
+
+        // THEN
+        XCTAssertEqual(returnedValue, 0, "Wrong returnedValue")
+
+        // THEN - UseCases
+        XCTAssertFalse(self.getColorsUseCase.executeWithThemeAndIntentCalled, "getColorsUseCase.executeWithThemeAndIntent shouldn't have been called")
+        XCTAssertFalse(self.getCornerRadiiUseCase.executeWithThemeAndShapeCalled, "getCornerRadiiUseCase.executeWithThemeAndShape shouldn't have been called")
+        XCTAssertFalse(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepCalled, "getStepValuesInBoundsUseCase.executeWithBoundsAndStep shouldn't have been called")
+        XCTAssertFalse(self.getClosestValueUseCase.executeWithValueAndValuesCalled, "getClosestValueUseCase.executeWithValueAndValues shouldn't have been called")
+
+        // THEN - Publishers
+        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius shouldn't have been called")
+    }
+
+    // MARK: Reset Bounds If Needed
+    func test_resetBoundsIfNeeded_is_not_needed() {
+        // GIVEN - Inits from setUp()
+        let expectedStepValues: [Float] = [0, 0.3, 1]
+        self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepReturnValue = expectedStepValues
+        self.viewModel.step = 0.4 // Changing bounds will not change stepValues unless step isn't nil
+        self.resetUseCases() // Removes previous executes
+        self.publishers.reset() // Removes previous publishes
+
+        // WHEN
+        self.viewModel.resetBoundsIfNeeded()
+
+        // THEN
+        XCTAssertEqual(self.viewModel.bounds, 0...1, "Wrong bounds")
+        XCTAssertEqual(self.viewModel.stepValues, expectedStepValues, "Wrong stepValues")
+
+        // THEN - UseCases
+        XCTAssertFalse(self.getColorsUseCase.executeWithThemeAndIntentCalled, "getColorsUseCase.executeWithThemeAndIntent shouldn't have been called")
+        XCTAssertFalse(self.getCornerRadiiUseCase.executeWithThemeAndShapeCalled, "getCornerRadiiUseCase.executeWithThemeAndShape shouldn't have been called")
+        XCTAssertFalse(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepCalled, "getStepValuesInBoundsUseCase.executeWithBoundsAndStep shouldn't have been called")
+        XCTAssertFalse(self.getClosestValueUseCase.executeWithValueAndValuesCalled, "getClosestValueUseCase.executeWithValueAndValues shouldn't have been called")
+
+        // THEN - Publishers
+        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius shouldn't have been called")
+    }
+
+    func test_resetBoundsIfNeeded_is_needed() throws {
+        // GIVEN - Inits from setUp()
+        let expectedStepValues: [Float] = [0, 0.3, 0.6, 0.9]
+        self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepReturnValue = expectedStepValues
+        self.viewModel.step = 0.3 // Changing bounds will not change stepValues unless step isn't nil
+        self.resetUseCases() // Removes previous executes
+        self.publishers.reset() // Removes previous publishes
+
+        // WHEN
+        self.viewModel.resetBoundsIfNeeded()
+
+        // THEN
+        XCTAssertEqual(self.viewModel.bounds, 0...0.9, "Wrong bounds")
+        XCTAssertEqual(self.viewModel.stepValues, expectedStepValues, "Wrong stepValues")
+
+        // THEN - UseCases
+        XCTAssertFalse(self.getColorsUseCase.executeWithThemeAndIntentCalled, "getColorsUseCase.executeWithThemeAndIntent shouldn't have been called")
+        XCTAssertFalse(self.getCornerRadiiUseCase.executeWithThemeAndShapeCalled, "getCornerRadiiUseCase.executeWithThemeAndShape shouldn't have been called")
+        XCTAssertFalse(self.getClosestValueUseCase.executeWithValueAndValuesCalled, "getClosestValueUseCase.executeWithValueAndValues shouldn't have been called")
+        XCTAssertEqual(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepCallsCount, 1, "getStepValuesInBoundsUseCase.executeWithBoundsAndStep should have been called once")
+        let receivedArguments = try XCTUnwrap(self.getStepValuesInBoundsUseCase.executeWithBoundsAndStepReceivedArguments, "Couldn't unwrap receivedArguments")
+        XCTAssertEqual(receivedArguments.step, 0.3, "Wrong received step")
+        XCTAssertEqual(receivedArguments.bounds, 0...0.9, "Wrong received bounds")
+
+        // THEN - Publishers
+        XCTAssertFalse(self.publishers.dim.sinkCalled, "$dim shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleColor.sinkCalled, "$handleColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.handleActiveIndicatorColor.sinkCalled, "$handleActiveIndicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackColor.sinkCalled, "$trackColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorColor.sinkCalled, "$indicatorColor shouldn't have been called")
+        XCTAssertFalse(self.publishers.trackRadius.sinkCalled, "$trackRadius shouldn't have been called")
+        XCTAssertFalse(self.publishers.indicatorRadius.sinkCalled, "$indicatorRadius shouldn't have been called")
     }
 }
